[HACK] Makefile: build: compile empty built-in object with cc
Building empty built-in object with ar creates an empty
file (which contains `!<arch>` for some reason???) but
if (only) one or more of these empty object is used to link
and create another built-in object (in a parent module)
then the linker uses default object headers/flags because
none of those object contains any valid object header,
and then this leads to a mismatch of eabi version because
other modules in the kernel might not have been compiled
with those default object flags/headers.
Using CC to build the empty built-in object fixes it as it
compiles a totally empty c file that has valid object headers
that are also used to build other modules in the kernel.

Change-Id: I3e6410ed8ff71d5fd19fddd82dc4042c408ffecb
Signed-off-by: parthibx24 <e.inxpired@gmail.com>